Add terminology model fetch
